# lib/sha256: Remove it