--- a/lib/build/build.go
+++ b/lib/build/build.go
@@ -40,7 +40,6 @@ var (
 
 	envTags = []string{
 		"STGUIASSETS",
-		"STHASHING",
 		"STNORESTART",
 		"STNOUPGRADE",
 	}