--- a/man/syncthing.1
+++ b/man/syncthing.1
@@ -555,11 +555,6 @@ passing \fI\%\-\-gui\-apikey\fP\&.
 Directory to load GUI assets from. Overrides compiled in assets. Useful for
 developing webgui, commonly use \fBSTGUIASSETS=gui bin/syncthing\fP\&.
 .TP
-.B STHASHING
-Specify which hashing package to use. Defaults to automatic based on
-performance. Specify “minio” (compatibility) or “standard” for the default
-Go implementation.
-.TP
 .B STHEAPPROFILE
 Write heap profiles to \fBheap\-$pid\-$timestamp.pprof\fP each time heap usage
 increases.