… go.sum
… lib/scanner/blocks.go
… lib/sha256/sha256.go