@imsodin 2024-08-10
That go ticket only talks about one intel instructions, while sha256-simd talks about multiple and also ARM specific things. Are we sure we aren't losing optimisations for some hardware by dropping this?