@calmh 2024-08-10
At least on macOS ARM64 it's the same.

```
[I6KAH] 2024/08/10 11:36:30 INFO: Single thread SHA256 performance is 2504 MB/s using crypto/sha256 (2494 MB/s using minio/sha256-simd).
```