@AudriusButkevicius 2024-08-10
Go version for amd64 doesn't seem to use native instructions, where as minio does.
This is only likely be visible on more modern cpus.