@AudriusButkevicius 2024-08-10
```
[2ERMY] 2024/08/10 12:37:04 INFO: Single thread SHA256 performance is 1740 MB/s using crypto/sha256 (1740 MB/s using minio/sha256-simd).
```
so we're good