@yonas 2024-09-04
@Gusted Would it be possible for you to add [BLAKE-3](https://www.infoq.com/news/2020/01/blake3-fast-crypto-hash/) to your comparison?